add lab work no 1
Print the strings in the following order:

   I

   know

   Java.

Calculate the result of the expression:

   (46 + 10) * (10 / 3)

   (29)  (4)  (-15)

In the variable number, there is an integer 10500. In the variable result, calculate the following expression: (number / 10) / 10. Print the result to the console.

Given three numbers: 3.6, 4.1, 5.9. In the variable result, calculate the product of these numbers.

In this task, you need to read integers from standard input and then output them. Each integer should be printed on a new line. Input format:

   42

   100

   125

For the integer b, perform the following conditional actions:

   ● If b is odd, print “Odd”

   ● If b is even, print “Even”

   ● If b is even and greater than 100, print “Out of range”

   ! Use the Scanner class: int b = scanner.nextInt();